+++ index.html
@@ -2,13 +2,19 @@
 <html>
 <head>
   <meta charset="utf-8">
-  <link rel="stylesheet" href="/lib/main.css" media="screen" title="no title">
+  <link rel="stylesheet" href="./lib/main.css" media="screen" title="no title">
   <title>Game Time</title>
 
 </head>
 <body>
   <canvas id="canvas" width="1250" height="650"></canvas>
-  <script src="main.bundle.js"></script>
-
 </body>
+<!-- <script src="main.bundle.js"></script> -->
+
+
+<script type="text/javascript" src="./lib/collision.js"></script>
+<script type="text/javascript" src="./lib/platform.js"></script>
+<script type="text/javascript" src="./lib/player.js"></script>
+<script type="text/javascript" src="./lib/game.js"></script>
+
 </html>
